refactor: corrected the redirect url, loading during the login process, logout flow
