+++ android/app/build.gradle
@@ -136,10 +136,9 @@ android {
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
         versionName "1.0"
-         manifestPlaceholders = [
-             // TODO get it from env file
-            appAuthRedirectScheme: "x-realdevsquad-rdsapp"
-    ]
+        manifestPlaceholders = [
+            appAuthRedirectScheme: 'com.rdsapp.auth'
+        ]
     }
     splits {
         abi {
